refactor: move drag-and-drop functionality to separate hook
